@@ -1,13 +1,12 @@
 apply plugin: 'com.android.library'
-apply from: 'build_library_ext.gradle'
+apply plugin: 'maven-publish'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.0"
+    compileSdkVersion 33
 
     defaultConfig {
         minSdkVersion 16
-        targetSdkVersion 23
+        targetSdkVersion 33
     }
     buildTypes {
         release {
@@ -16,13 +15,24 @@ android {
         }
     }
 }
-
+afterEvaluate {
+    publishing {
+        publications {
+            // Creates a Maven publication called "release".
+            release(MavenPublication) {
+                from components.release
+                groupId = 'com.github.whosmyqueen'
+                artifactId = 'android-drawing-view'
+                version = '1.0.0'
+            }
+        }
+    }
+}
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:23.0.0'
-
-    compile 'com.github.vilyever:AndroidJsonModel:1.2.0'
-    compile 'com.github.vilyever:AndroidFileReadWrite:1.1.0'
-    compile 'com.github.vilyever:AndroidUnitConversion:1.1.0'
-    compile 'com.github.vilyever:AndroidResource:1.2.1'
+    api fileTree(dir: 'libs', include: ['*.jar'])
+    api "androidx.appcompat:appcompat:1.2.0"
+    api 'com.github.vilyever:AndroidJsonModel:1.2.0'
+    api 'com.github.vilyever:AndroidFileReadWrite:1.1.0'
+    api 'com.github.vilyever:AndroidUnitConversion:1.1.0'
+    api 'com.github.vilyever:AndroidResource:1.2.1'
 }
