@@ -4,7 +4,8 @@ import android.graphics.Canvas;
 import android.graphics.Color;
 import android.graphics.Path;
 import android.graphics.RectF;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
 
 import com.vilyever.drawingview.R;
 import com.vilyever.drawingview.model.DrawingPath;
@@ -21,7 +22,7 @@ import com.vilyever.resource.Resource;
 public class CircleBrush extends ShapeBrush {
     final CircleBrush self = this;
 
-    
+
     /* #Constructors */
     public CircleBrush() {
 
