@@ -3,7 +3,8 @@ package com.vilyever.drawingview.brush.drawing;
 import android.graphics.BlurMaskFilter;
 import android.graphics.Canvas;
 import android.graphics.Color;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
 
 import com.vilyever.drawingview.R;
 import com.vilyever.drawingview.model.DrawingPath;
@@ -20,7 +21,7 @@ import com.vilyever.resource.Resource;
 public class GlowPenBrush extends PenBrush {
     final GlowPenBrush self = this;
 
-    
+
     /* Constructors */
     public GlowPenBrush() {
     }
@@ -28,25 +29,27 @@ public class GlowPenBrush extends PenBrush {
     public GlowPenBrush(float size, int color) {
         super(size, color);
     }
-    
-    
+
+
     /* Public Methods */
     public static GlowPenBrush defaultBrush() {
         return new GlowPenBrush(Resource.getDimensionPixelSize(R.dimen.drawingViewBrushDefaultSize), Color.RED);
     }
-    
-    
+
+
     /* Properties */
     private boolean onBlurDraw;
+
     private <T extends GlowPenBrush> T setOnBlurDraw(boolean onBlurDraw) {
         this.onBlurDraw = onBlurDraw;
         return (T) this;
     }
+
     public boolean isOnBlurDraw() {
         return this.onBlurDraw;
     }
-    
-    
+
+
     /* Overrides */
     @Override
     protected void updatePaint() {
@@ -55,8 +58,7 @@ public class GlowPenBrush extends PenBrush {
         if (isOnBlurDraw()) {
             getPaint().setStrokeWidth(getSize() * 2.0f);
             getPaint().setMaskFilter(new BlurMaskFilter(getSize(), BlurMaskFilter.Blur.NORMAL));
-        }
-        else {
+        } else {
             getPaint().setColor(Color.WHITE);
             getPaint().setMaskFilter(null);
         }
@@ -70,10 +72,10 @@ public class GlowPenBrush extends PenBrush {
         setOnBlurDraw(false);
         return super.drawPath(canvas, drawingPath, state);
     }
-     
+
     /* Delegates */
-     
-     
+
+
     /* Private Methods */
-    
+
 }
