@@ -4,7 +4,8 @@ import android.graphics.Canvas;
 import android.graphics.Color;
 import android.graphics.Path;
 import android.graphics.RectF;
-import android.support.annotation.NonNull;
+
+import androidx.annotation.NonNull;
 
 import com.vilyever.drawingview.R;
 import com.vilyever.drawingview.model.DrawingPath;
@@ -106,5 +107,5 @@ public class RoundedRectangleBrush extends ShapeBrush {
 
         return Frame.EmptyFrame();
     }
-    
+
 }
